Rename repos to stores
