+++ packages/db-drizzle/package.json
@@ -5,7 +5,7 @@
 	"type": "module",
 	"exports": {
 		"./db": "./src/db/db.ts",
-		"./repositories/*": "./src/repositories/*",
+		"./stores/*": "./src/stores/*",
 		"./schema": "./src/db/schema.ts",
 		"./schema-type": "./src/db/schema-type-helpers.ts"
 	},
